Fixed problems in GameSettings
The game now gets a default level of 1 if the user doesnt set anything. Previously the variables holding this would be changed even if the game wasnt set correctly. This is now fixed by adding holder variables which resets the variables which hold the textfield values in case the textfieldvalues were incorrect.